Changed NavbarList component name to Navbar
