+++ src/assets/styles/main.scss
@@ -14,6 +14,7 @@
 @import 'components/contact-form';
 @import 'components/divisor';
 @import 'components/modal';
+@import 'components/navbar';
 @import 'components/opacity-button';
 @import 'components/preview-redirector';
