+++ src/components/index.ts
@@ -2,5 +2,6 @@ export { default as Button } from './Button';
 export { default as ContactForm } from './ContactForm';
 export { default as Divisor } from './Divisor';
 export { default as Modal } from './Modal';
-export { default as NavbarList } from './NavbarList';
+export { default as Navbar } from './Navbar';
+export { default as OpacityButton } from './OpacityButton';
 export { default as PreviewButton } from './PreviewButton';
